Use WIN32OLE to resolve networkadapter_product
Previously the networkadapter_product fact used the Facter::Util::WMI
API to resolve the value, but that API is no longer available in Facter
3.

This commit modifies the fact to use WIN32OLE instead, as is done with
other WMI based facts in this module. The WIN32OLE API is provided by
Ruby so it isn't sensitive to facter versions.

Fixes #3.